backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
pandas>=2.1.0
scipy>=1.11.0
requests>=2.31.0
PyYAML>=6.0

